// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/meal-plan-day-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MealPlanMealHttpResponse } from './meal-plan-meal-http-response';

export interface MealPlanDayHttpResponse {
    dayNumber?: number;
    id?: string;
    meals?: Array<MealPlanMealHttpResponse>;
}