// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/meal-plan-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MealPlanDayHttpResponse } from './meal-plan-day-http-response';

export interface MealPlanHttpResponse {
    days?: Array<MealPlanDayHttpResponse>;
    description?: string | null;
    dietType?: string;
    durationDays?: number;
    id?: string;
    isCurated?: boolean;
    name?: string;
    targetCaloriesPerDay?: number | null;
}